chore: Move projects' components to a dedicated folder
