+++ src/components/index.ts
@@ -1,6 +1,6 @@
 export { PageSelector } from "./PageSelector";
-export { EditProjectDialog } from "./EditProjectDialog";
-export { CreateProjectDialog } from "./CreateProjectDialog";
+export { EditProjectDialog } from "./projects/EditProjectDialog";
+export { CreateProjectDialog } from "./projects/CreateProjectDialog";
 export { ProgressBar } from "./ProgressBar";
 export { AddPluginDialog } from "@/components/plugins/AddPluginDialog";
 export { EditPluginDialog } from "@/components/plugins/EditPluginDialog";
